Move test to monitoring package
